Fix Encapsulation Problem
In file: /workspace/source/src/main/java/baritone/behavior/Behavior.java, class: Behavior has a field that is declared as public but it may allow unwarranted access. The access to the field should be restricted and should only be through accessor methods. iCR suggested changes in 7 files to resolve the problem. For more information, read <hyperlink1>