Fix Encapsulation Problem
In file: PathNode.java, class: PathNode has a field that is declared as public but it may allow unwarranted access. The access to the field should be restricted and should only be through accessor methods. iCR suggested changes in 2 files to resolve the problem. For more information, read CERT Secure_coding_standard_OBJ01-J.