Fix Use Better Directory Creation API
In file: ProguardTask.java, there is a method acquireDependencies that uses an API to create a directory which can fail if the parent directory does not exist. iCR suggested to use a better directory creation API such as File.mkdirs instead of File.mkdir and Path.createDirectories instead of Path.createDirectory as the former APIs can create the non-existent parent directory as needed.